requirement.txt
Flask
Flask-CORS
pymongo
pandas
jsonify
python-dotenv
google-generativeai
google-api-python-client
googletrans 